put callbacks in folder and renamed figures folder
