+++ webviz_subsurface/plugins/_rft_plotter/_callbacks/__init__.py
@@ -1,0 +1,1 @@
+from .callbacks import plugin_callbacks
