+++ webviz_subsurface/plugins/_rft_plotter/_figures/_formation_figure.py
