+++ webviz_subsurface/plugins/_rft_plotter/_figures/_map_figure.py
