+++ webviz_subsurface/plugins/_rft_plotter/_figures/_misfit_figure.py
